refactor test code, add test for audit log
