+++ integration/integration.cabal
@@ -118,7 +118,6 @@ library
     Test.Brig
     Test.Cargohold.API
     Test.Cargohold.API.Federation
-    Test.Cargohold.API.Util
     Test.Cargohold.API.V3
     Test.Cargohold.Metrics
     Test.Cells
