+++ integration/test/Test/AssetUpload.hs
@@ -8,13 +8,13 @@ import Testlib.Prelude
 testAssetUploadUnverifiedUser :: (HasCallStack) => App ()
 testAssetUploadUnverifiedUser = do
   user <- randomUser OwnDomain $ def {activate = False}
-  bindResponse (uploadAsset user) $ \resp -> do
+  bindResponse (uploadSomeAsset user) $ \resp -> do
     resp.status `shouldMatchInt` 403
 
 testAssetUploadVerifiedUser :: (HasCallStack) => App ()
 testAssetUploadVerifiedUser = do
   user <- randomUser OwnDomain def
-  bindResponse (uploadAsset user) $ \resp -> do
+  bindResponse (uploadSomeAsset user) $ \resp -> do
     resp.status `shouldMatchInt` 201
 
 testAssetUploadUnknownUser :: (HasCallStack) => App ()
@@ -30,5 +30,5 @@ testAssetUploadUnknownUser = do
                   "id" .= uid
                 ]
           ]
-  bindResponse (uploadAsset user) $ \resp -> do
+  bindResponse (uploadSomeAsset user) $ \resp -> do
     resp.status `shouldMatchInt` 403
