+++ integration/test/Test/Cargohold/API.hs
@@ -34,8 +34,7 @@ import Data.Time.Format.ISO8601 (formatParseM, iso8601Format)
 import Network.HTTP.Client (parseUrlThrow)
 import qualified Network.HTTP.Client as HTTP
 import qualified Network.HTTP.Types as HTTP
-import SetupHelpers (randomId, randomUser)
-import Test.Cargohold.API.Util
+import SetupHelpers (createTeam, randomId, randomUser)
 import Testlib.Prelude
 import UnliftIO.Concurrent
 
@@ -56,7 +55,7 @@ testSimpleRoundtrip = do
       uid2 <- randomUser OwnDomain def
       -- Initial upload
       let bdy = (applicationText, cs "Hello World")
-      r1 <- uploadSimple uid sets bdy
+      r1 <- uploadSimpleV3 uid sets bdy
       r1.status `shouldMatchInt` 201
       loc <- maybe (error "Could not find the Location header") (pure . cs @_ @String) $ lookup (mk $ cs "Location") r1.headers
       (tok, expires) <-
@@ -124,7 +123,7 @@ testSimpleTokens = do
   -- Initial upload
   let sets = object ["public" .= False, "retention" .= "volatile"]
       bdy = (applicationText, cs "Hello World")
-  r1 <- uploadSimple uid sets bdy
+  r1 <- uploadSimpleV3 uid sets bdy
   r1.status `shouldMatchInt` 201
   loc <-
     maybe
@@ -208,7 +207,7 @@ testSimpleS3ClosedConnectionReuse = go >> wait >> go
       uid <- randomUser OwnDomain def
       let sets = object $ defAssetSettings' <> ["retention" .= "volatile"]
       let part2 = (MIME.Text $ cs "plain", cs $ replicate 100000 'c')
-      uploadSimple uid sets part2 >>= \r -> r.status `shouldMatchInt` 201
+      uploadSimpleV3 uid sets part2 >>= \r -> r.status `shouldMatchInt` 201
 
 testDownloadURLOverride :: (HasCallStack) => App ()
 testDownloadURLOverride = do
@@ -223,7 +222,7 @@ testDownloadURLOverride = do
     uid <- randomUser d def
     -- Upload, should work, shouldn't try to use the S3DownloadEndpoint
     let bdy = (applicationText, cs "Hello World")
-    uploadRes <- uploadSimple uid defAssetSettings bdy
+    uploadRes <- uploadSimpleV3 uid defAssetSettings bdy
     uploadRes.status `shouldMatchInt` 201
     let loc = decodeHeaderOrFail (mk $ cs "Location") uploadRes :: String
     (_key, tok, _expires) <-
@@ -255,7 +254,7 @@ testUploadCompatibility :: (HasCallStack) => App ()
 testUploadCompatibility = do
   uid <- randomUser OwnDomain def
   -- Initial upload
-  r1 <- uploadRaw uid exampleMultipart
+  r1 <- uploadRawV3 uid exampleMultipart
   r1.status `shouldMatchInt` 201
   let locHeader = mk $ cs "Location"
       loc = decodeHeaderOrFail @String locHeader r1
@@ -326,7 +325,7 @@ remoteDownload :: (HasCallStack, ConvertibleStrings a L8.ByteString, Convertible
 remoteDownload content = do
   uid1 <- randomUser OwnDomain def
   uid2 <- randomUser OtherDomain def
-  r1 <- uploadSimple uid1 settings (applicationOctetStream, cs content)
+  r1 <- uploadSimpleV3 uid1 settings (applicationOctetStream, cs content)
   r1.status `shouldMatchInt` 201
   let locHeader = mk $ cs "Location"
       loc = decodeHeaderOrFail @String locHeader r1
@@ -338,3 +337,23 @@ remoteDownload content = do
   cs @_ @String r2.body `shouldMatch` Just (cs content :: String)
   where
     settings = object ["public" .= True]
+
+------------------------------------------------------------------------
+-- Asset Audit Log
+
+testAssetAuditLogUpload :: (HasCallStack) => App ()
+testAssetAuditLogUpload = do
+  (owner, _tid, _members) <- createTeam OwnDomain 1
+  convUuid <- randomId
+  domain <- owner %. "qualified_id.domain" & asString
+  let settings =
+        object
+          [ "public" .= False,
+            "convId" .= object ["id" .= convUuid, "domain" .= domain],
+            "filename" .= "virus.js",
+            "filetype" .= "application/javascript"
+          ]
+      body = (applicationText, cs "Hello Audit Log")
+  -- Upload should succeed and return 201 even with extra metadata fields
+  r <- uploadSimple owner settings body
+  r.status `shouldMatchInt` 201
